# DOCS/.architecture/openapi.yaml
...
openapi: 3.1.0
info:
  title: Credit Dashboard SUT API
  version: 0.8.1
  license:
    name: MIT
    identifier: MIT
  summary: Fictional credit-report service used as a system under test. All data is synthetic.
  description: |
    Phase 0 draft contract. Business rule IDs (BR-xx) refer to the API specification;
    profile rule IDs (PR-xx) refer to the My Profile UI feature spec.
    Test-control paths are tagged `test-control` and exist only when TEST_CONTROL=true.
    Lint ruleset: `redocly.yaml` at the repository root.
servers:
  - url: http://localhost:4000/api/v1
    description: Local service
  - url: http://localhost:4010
    description: Prism mock. Prism serves the paths without the /api/v1 base path (DR-039)
security:
  - bearerAuth: []
...
paths:
...
  /__test/reset:
    post:
      tags: [test-control]
      operationId: testReset
      summary: Reset fixtures, flags, clock and latency
      security: [ { testControlKey: [] } ]
      responses:
        '204': { description: 'Fixtures reloaded; flags, clock and latency cleared' }
        '404': { $ref: '#/components/responses/TestControlDisabled' }
  /__test/users/{username}/persona:
    put:
      tags: [test-control]
      operationId: testBindPersona
      summary: Bind a test user to a persona, optionally with data overrides
      description: >-
        Binds the user to the persona. Optional `overrides` arrange data no persona holds (DR-020): each list
        named for a bureau replaces that bureau's list, for this user only, until rebound or reset. Overrides are
        source data in the fixture format (API spec section 9.1); derived values are recomputed from them. A body
        that fails the schema is a 400; overrides that break a business rule (for example a stored utilisation
        that does not match its balance and limit, BR-03) are a 422 (`/problems/rule-violation/overrides-inconsistent`);
        an unknown bureau ID is a 404.
      security: [ { testControlKey: [] } ]
      parameters:
        - { name: username, in: path, required: true, schema: { type: string, example: alex } }
      requestBody:
        required: true
        content:
          application/json:
            examples:
              plain:
                summary: Bind only
                value:
                  persona: drilldown
              withOverrides:
                summary: One credit card, for a BR-03 outline row
                value:
                  persona: drilldown
                  overrides:
                    bureaux:
                    - id: bureau-a
                      accounts:
                      - id: acc_ovcc01
                        type: creditcard
                        provider: Lender X
                        logoUrl: null
                        maskedNumber: '*0001'
                        balance: { amountMinor: 4950, currency: GBP }
                        limit: { amountMinor: 10000, currency: GBP }
                        utilisation: 50
                        includedInTotals: true
                        status: normal
                        openedDate: '2024-01-15'
                        closedDate: null
                        updateFrequency: monthly
                        lastUpdated: '2026-09-30'
                        utilisationRaw: 50
                        details: { apr: null, interestRate: null, promoPeriodMonths: null, minPayment: null, paymentMethod: null }
                        closed: false
                        missedMonths: []
                        balanceHistory:
                        - { month: '2026-09', balance: { amountMinor: 4950, currency: GBP }, limit: { amountMinor: 10000, currency: GBP } }
            schema:
              type: object
              additionalProperties: false
              required: [persona]
              properties:
                persona: { $ref: '#/components/schemas/Persona' }
                overrides: { $ref: '#/components/schemas/PersonaOverrides' }
      responses:
        '204': { description: Bound }
        '400': { $ref: '#/components/responses/Validation' }
        '404': { $ref: '#/components/responses/TestControlDisabled' }
        '422': { $ref: '#/components/responses/OverridesInconsistent' }
  /__test/bugs:
    put:
      tags: [test-control]
      operationId: testSetBugs
      summary: Set the active bug flags
      security: [ { testControlKey: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            example:
              flags:
              - util-mismatch
              - minor-units-label
            schema:
              type: object
              required: [flags]
              properties:
                flags:
                  type: array
                  uniqueItems: true
                  items: { $ref: '#/components/schemas/BugFlag' }
      responses:
        '204': { description: Flags set }
        '400': { $ref: '#/components/responses/Validation' }
        '404': { $ref: '#/components/responses/TestControlDisabled' }
  /__test/clock:
    put:
      tags: [test-control]
      operationId: testSetClock
      summary: Freeze server time
      security: [ { testControlKey: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            example:
              now: '2026-10-03T09:00:00Z'
            schema:
              type: object
              required: [now]
              properties: { now: { type: string, format: date-time } }
      responses:
        '204': { description: Clock frozen }
        '400': { $ref: '#/components/responses/Validation' }
        '404': { $ref: '#/components/responses/TestControlDisabled' }
  /__test/latency:
    put:
      tags: [test-control]
      operationId: testSetLatency
      summary: Set artificial latency
      security: [ { testControlKey: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            example:
              fixedMs: 3000
            schema:
              type: object
              properties:
                fixedMs: { type: integer, minimum: 0, maximum: 10000 }
                minMs: { type: integer, minimum: 0 }
                maxMs: { type: integer, maximum: 10000 }
      responses:
        '204': { description: Latency set }
        '400': { $ref: '#/components/responses/Validation' }
        '404': { $ref: '#/components/responses/TestControlDisabled' }
  /__test/state:
    get:
      tags: [test-control]
      operationId: testGetState
      summary: Read the current test-control state
      security: [ { testControlKey: [] } ]
      responses:
        '200':
          description: Current control state
          content:
            application/json:
              example:
                personas:
                  alex: excellent
                overridden:
                  alex: false
                flags: []
                now: null
                latency: {}
              schema:
                type: object
                properties:
                  personas: { type: object, additionalProperties: { $ref: '#/components/schemas/Persona' } }
                  overridden:
                    description: Whether each bound user has overrides in force (DR-020)
                    type: object
                    additionalProperties: { type: boolean }
                  flags: { type: array, items: { $ref: '#/components/schemas/BugFlag' } }
                  now: { type: [string, 'null'], format: date-time }
                  latency: { type: object }
        '404': { $ref: '#/components/responses/TestControlDisabled' }
  /__test/verify-email:
    post:
      tags: [test-control]
      operationId: testVerifyEmail
      summary: Mark a test user's email verified without the link
      security: [ { testControlKey: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            example: { username: alex }
            schema:
              type: object
              additionalProperties: false
              required: [username]
              properties:
                username: { type: string }
      responses:
        '204': { description: Email marked verified }
        '400': { $ref: '#/components/responses/Validation' }
        '404': { $ref: '#/components/responses/TestControlDisabled' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    testControlKey: { type: apiKey, in: header, name: X-Test-Control-Key }
...
  responses:
    Validation:
      description: Request failed validation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: /problems/validation
            title: Request failed validation
            status: 400
            detail: One or more fields are invalid.
            instance: /api/v1/reports/bureau-a/score/history
            errors:
            - field: range
              message: must be one of 3m, 6m, 1y
...
    OverridesInconsistent:
      description: Test-control overrides break a business rule (DR-020)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: /problems/rule-violation/overrides-inconsistent
            title: Overrides break a business rule
            status: 422
            detail: acc_ovcc01 utilisation 40 does not match its balance and limit (BR-03).
            instance: /api/v1/__test/users/alex/persona
...
    TestControlDisabled:
      description: Test control is disabled (TEST_CONTROL is not true), or the target does not exist (DR-008)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: /problems/not-found
            title: Not found
            status: 404
            detail: Test control is not enabled.
            instance: /api/v1/__test/reset
  schemas:
    Money:
      type: object
      required: [amountMinor, currency]
      properties:
        amountMinor: { type: integer, description: Minor units (pence). Negative means a credit balance. }
        currency: { type: string, enum: [GBP] }
    Utilisation:
      type: [integer, 'null']
      minimum: 0
      description: >-
        Balance over limit, x 100, rounded half up (BR-03). No upper bound: above 100 means over
        the limit (DR-013). Floored at 0 for a negative balance (BR-06). Null when the limit is zero or absent.
    Sentiment: { type: string, enum: [positive, neutral, negative] }
    AccountType: { type: string, enum: [creditcard, loan, mortgage, currentaccount, telecomsandutilities, lineofcredit] }
    AccountStatus: { type: string, enum: [normal, arrears, default, settled, closed] }
...
    Persona: { type: string, enum: [excellent, struggling, thin-file, boundary, drilldown, error, slow] }
    BugFlag:
      description: >-
        A defect that test control can switch on (DR-008). The API flags are in API specification section 10; the UI and
        profile flags are in the UI specification section 8 and the My Profile feature specification. A name outside this list is
        a 400, so a typo cannot silently do nothing.
      type: string
      enum:
        - excluded-in-total
        - rounding-down
        - history-carry-forward
        - idor
        - problem-json-missing
        - mask-format
        - currency-float
        - duplicate-nav
        - util-mismatch
        - currency-format
        - minor-units-label
        - placeholder-aria
        - chart-no-values
        - range-stale
        - toggle-label
        - like-both
        - plural
        - button-href
        - double-render
        - ph-a11y
        - open-redirect
        - truncate-summary
        - negative-balance
        - nested-button
        - decorative-alt
        - preferred-on-report
        - preferred-no-trim
        - email-stays-verified
        - pii-in-title
        - lost-edit
    MaskedNumber: { type: string, pattern: '^\*[A-Z0-9]{4}$' }
...
    Change:
      type: object
      required: [id, title, sentiment, impact, date]
      properties:
        id: { type: string }
        title: { type: string }
        sentiment: { $ref: '#/components/schemas/Sentiment' }
        impact: { type: string, enum: [low, medium, high] }
        date: { type: string, format: date }
...
    AccountSummary:
      type: object
      required: [id, type, provider, maskedNumber, balance, includedInTotals, status]
      properties:
        id: { type: string }
        type: { $ref: '#/components/schemas/AccountType' }
        provider: { type: string }
        logoUrl: { type: [string, 'null'], format: uri }
        maskedNumber: { $ref: '#/components/schemas/MaskedNumber' }
        balance: { $ref: '#/components/schemas/Money' }
        limit: { oneOf: [ { $ref: '#/components/schemas/Money' }, { type: 'null' } ] }
        utilisation: { $ref: '#/components/schemas/Utilisation' }
        includedInTotals: { type: boolean }
        status: { $ref: '#/components/schemas/AccountStatus' }
...
    AccountDetails:
      type: object
      properties:
        apr: { type: [number, 'null'], minimum: 0, maximum: 100 }
        interestRate: { type: [number, 'null'], minimum: 0, maximum: 100 }
        promoPeriodMonths: { type: [integer, 'null'], minimum: 0, maximum: 60 }
        minPayment:
          type: [object, 'null']
          properties:
            amount: { $ref: '#/components/schemas/Money' }
            percent: { type: number, minimum: 0, maximum: 100 }
        paymentMethod: { type: [string, 'null'], enum: [direct-debit, manual, null] }
    Account:
      allOf:
        - { $ref: '#/components/schemas/AccountSummary' }
        - type: object
          required: [openedDate, updateFrequency, lastUpdated, details, closed]
          properties:
            openedDate: { type: string, format: date }
            closedDate: { type: [string, 'null'], format: date }
            updateFrequency: { type: string, enum: [monthly, quarterly] }
            lastUpdated: { type: string, format: date }
            utilisationRaw: { type: [integer, 'null'], description: Unfloored value (BR-06) }
            details: { $ref: '#/components/schemas/AccountDetails' }
            closed: { type: boolean }
...
    BalancePoint:
      type: object
      required: [month, balance]
      properties:
        month: { type: string, pattern: '^\d{4}-\d{2}$' }
        balance: { $ref: '#/components/schemas/Money' }
        limit: { oneOf: [ { $ref: '#/components/schemas/Money' }, { type: 'null' } ] }
    FixtureAccount:
      description: >-
        An account as source data (API spec section 9.1): the contract Account plus the inputs the service derives
        from. Used by persona fixtures and by test-control overrides (DR-020).
      allOf:
        - { $ref: '#/components/schemas/Account' }
        - type: object
          required: [missedMonths, balanceHistory]
          properties:
            sourceMask: { type: string, minLength: 1, description: BR-09 input as the source supplies it }
            missedMonths:
              description: BR-12 source
              type: array
              uniqueItems: true
              items: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }
            balanceHistory:
              description: Oldest first; also the BR-07 trend source
              type: array
              maxItems: 6
              items: { $ref: '#/components/schemas/BalancePoint' }
    PersonaOverrides:
      description: >-
        Test-control data for one binding (DR-020). Each list named for a bureau replaces that list; lists not
        named are kept from the persona.
      type: object
      additionalProperties: false
      required: [bureaux]
      properties:
        bureaux:
          type: array
          minItems: 1
          items:
            type: object
            additionalProperties: false
            required: [id]
            minProperties: 2
            properties:
              id: { type: string }
              accounts: { type: array, items: { $ref: '#/components/schemas/FixtureAccount' } }
              changes: { type: array, items: { $ref: '#/components/schemas/Change' } }
              searches: { type: array, items: { $ref: '#/components/schemas/Search' } }
...
    Search:
      type: object
      required: [id, kind, organisation, date]
      properties:
        id: { type: string }
        kind: { type: string, enum: [hard, soft] }
        organisation: { type: string }
        purpose: { type: string }
        date: { type: string, format: date }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          description: >-
            The problem type (API specification section 8). A 422 names its rule outcome under
            /problems/rule-violation/ (DR-048); clients and tests branch on type, never on title or detail.
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        attemptsRemaining:
          type: integer
          minimum: 0
          maximum: 3
          description: code-wrong only (PR-11); code-invalid carries none
        errors:
          type: array
          items:
            type: object
            properties: { field: { type: string }, message: { type: string } }